refactore/fix: fix brand type, test and service
