+++ web/src/types/brand.ts
@@ -1,3 +1,5 @@
+// import type { Event } from './event';
+
 export interface Brand {
   id: number;
   name: string;
@@ -8,14 +10,20 @@ export interface Brand {
   secondary_color?: string;
 }
 
+// will reuse the Event from event type
 export interface BrandEvent {
   id: number;
   title: string;
   start_date: string;
+  end_date?: string;
+  location?: string;
+  status: string;
+  brand_id: number;
 }
 
 export interface BrandWithEvents extends Brand {
   events: BrandEvent[];
+  // events: Event[];
 }
 
 export interface CreateBrandRequest {
